调整doc-editor页面路由
