--- src/pages/doc-editor/doc-editor.jsx
@@ -1,28 +1,0 @@
-import React, { useState } from 'react';
-import { Document, Page } from 'react-pdf/dist/esm/entry.webpack';
-// import sample from '../../static/test.pdf'
-
-export default function DocEditor() {
-    const [numPages, setNumPages] = useState(null);
-    const [pageNumber, setPageNumber] = useState(1);
-
-    function onDocumentLoadSuccess({ numPages }) {
-        setNumPages(numPages);
-    }
-    function onDocumentLoadError(err) {
-        alert('Error while loading document! ' + err.message)
-    }
-
-    return (
-        <div>
-            <Document
-                file='/upload/docs/test.pdf'
-                onLoadSuccess={onDocumentLoadSuccess}
-                onLoadError={onDocumentLoadError}
-            >
-                <Page pageNumber={pageNumber} />
-            </Document>
-            <p>Page {pageNumber} of {numPages}</p>
-        </div>
-    );
-}
