+++ src/pages/upload/doc-editor.less
@@ -1,0 +1,20 @@
+.site-drawer-render-in-current-wrapper {
+    position: relative;
+    height: 100%;
+    padding: 48px;
+    overflow: hidden;
+    text-align: center;
+    background: #fafafa;
+    border: 1px solid #ebedf0;
+    border-radius: 2px;
+    margin-top: 20px;
+    .pdf-container{
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        .page-controls{
+            position: absolute;
+            display: flex;
+        }
+    }
+}
