+++ src/pages/upload/upload.less
